refactor(releases): move the global perspective label
